test: improve test quality and setup
- Fix #53: Removed fake assertions (assert True, assert isinstance object) and replaced exception swallowing with specific ProviderError catches.

- Fix #54: Moved test_advanced_glcs.py to unit tests as it mocks the LLM.

- Fix #56: Verified @pytest.mark.requires_api_key is properly registered.

- Fix #57: Created conftest.py with shared fixtures to reduce duplication.
